+++ lib/admin/template-parts.php
@@ -49,7 +49,7 @@ function mai_import_template_part_action() {
 
 	} else {
 		wp_die(
-			__( 'Template Parts failed to generate.', 'mai-engine' ),
+			__( 'Content Areas failed to generate.', 'mai-engine' ),
 			__( 'Error', 'mai-engine' ), [
 				'link_url'  => admin_url( 'edit.php?post_type=mai_template_part' ),
 				'link_text' => __( 'Go back.', 'mai-engine' ),
@@ -60,7 +60,7 @@ function mai_import_template_part_action() {
 
 add_action( 'admin_post_mai_generate_template_parts_action', 'mai_generate_template_parts_action' );
 /**
- * Listener for generating default template parts.
+ * Listener for generating default content areas.
  *
  * @since 2.10.0
  *
@@ -78,13 +78,13 @@ function mai_generate_template_parts_action() {
 
 		switch ( $count ) {
 			case 0:
-				$message = __( 'Sorry, no template parts are available.', 'mai-engine' );
+				$message = __( 'Sorry, no content areas are available.', 'mai-engine' );
 			break;
 			case 1:
-				$message = sprintf( '%s %s', $count, __( 'default template parts successfully created.', 'mai-engine' ) );
+				$message = sprintf( '%s %s', $count, __( 'default content areas successfully created.', 'mai-engine' ) );
 			break;
 			default:
-				$message = sprintf( '%s %s', $count, __( 'default template parts successfully created.', 'mai-engine' ) );
+				$message = sprintf( '%s %s', $count, __( 'default content areas successfully created.', 'mai-engine' ) );
 		}
 
 
@@ -97,7 +97,7 @@ function mai_generate_template_parts_action() {
 
 	} else {
 		wp_die(
-			__( 'Template Parts failed to generate.', 'mai-engine' ),
+			__( 'Content Areas failed to generate.', 'mai-engine' ),
 			__( 'Error', 'mai-engine' ), array(
 				'link_url'  => admin_url( 'edit.php?post_type=mai_template_part' ),
 				'link_text' => __( 'Go back.', 'mai-engine' ),
@@ -107,7 +107,7 @@ function mai_generate_template_parts_action() {
 
 add_action( 'load-edit.php', 'mai_template_parts_admin_notice' );
 /**
- * Adds admin notice to template parts.
+ * Adds admin notice to content areas.
  *
  * @since 2.6.0
  *
@@ -124,7 +124,7 @@ function mai_template_parts_admin_notice() {
 		printf(
 			'<div class="notice notice-success is-dismissible"><p>%s <a target="_blank" href="https://docs.bizbudding.com/docs/template-parts/">%s</a>.</p></div>',
 			__( 'View documentation for', 'mai-engine' ),
-			__( 'Template Parts', 'mai-engine' )
+			__( 'Content Areas', 'mai-engine' )
 		);
 	});
 
@@ -155,9 +155,9 @@ function mai_template_parts_admin_notice() {
 	add_action( 'admin_notices', function() use ( $available ) {
 
 		if ( 1 === $available ) {
-			$notice = sprintf( '%s %s', $available, __( 'default Template Part needs to be created.', 'mai-engine' ) );
+			$notice = sprintf( '%s %s', $available, __( 'default Content Area needs to be created.', 'mai-engine' ) );
 		} else {
-			$notice = sprintf( '%s %s', $available, __( 'default Template Parts need to be created.', 'mai-engine' ) );
+			$notice = sprintf( '%s %s', $available, __( 'default Content Areas need to be created.', 'mai-engine' ) );
 		}
 
 		$generate_url = add_query_arg( [ 'action' => 'mai_generate_template_parts_action' ], admin_url( 'admin-post.php' ) );
@@ -243,7 +243,7 @@ function mai_template_parts_import_row_action( $actions, $post ) {
 
 add_filter( 'display_post_states', 'mai_template_part_post_state', 10, 2 );
 /**
- * Display active template parts.
+ * Display active content areas.
  *
  * @since 2.0.0
  *
@@ -270,7 +270,7 @@ function mai_template_part_post_state( $states, $post ) {
 
 add_filter( 'manage_mai_template_part_posts_columns', 'mai_template_part_add_slug_column' );
 /**
- * Add slug column to Template Parts.
+ * Add slug column to Content Areas.
  * Inserts as second to last item.
  *
  * @since 2.0.0
@@ -337,7 +337,7 @@ function mai_template_parts_order( $query ) {
 
 add_action( 'current_screen', 'mai_widgets_template_parts_admin_notice' );
 /**
- * Adds admin notice for template parts to widgets screen.
+ * Adds admin notice for content areas to widgets screen.
  *
  * @since 2.6.0
  *
@@ -351,9 +351,9 @@ function mai_widgets_template_parts_admin_notice( $screen ) {
 	add_action( 'admin_notices', function() {
 		printf(
 			'<div class="notice notice-warning is-dismissible"><p>%s <a href="%s">%s</a>.</p></div>',
-			__( 'Mai Theme uses "Template Parts" (block-based widget areas).', 'mai-engine' ),
+			__( 'Mai Theme uses "Content Areas" (block-based widget areas).', 'mai-engine' ),
 			admin_url( 'edit.php?post_type=mai_template_part' ),
-			__( 'Edit template parts now', 'mai-engine' )
+			__( 'Edit content areas now', 'mai-engine' )
 		);
 	});
 }