+++ lib/init.php
@@ -236,15 +236,18 @@ function mai_genesis_theme_settings_menu_ops( $options ) {
  * @since 2.4.3
  * @since 2.6.0 Changed function name to avoid clash when switching from v1 to v2.
  * @since 2.6.0 Check mai_get_url() function exists. We saw this run too early and fail.
+ * @since 2.18.0 Added `get_site_icon_url` filter because it was throwing errors in admin for missing favicon.
  * @link  https://github.com/maithemewp/mai-engine/issues/361
  *
  * @return string
  */
+add_filter( 'get_site_icon_url', 'mai_load_default_favicon' );
 add_filter( 'genesis_pre_load_favicon', 'mai_load_default_favicon' );
 function mai_load_default_favicon( $favicon ) {
-	if ( function_exists( 'mai_get_url' ) ) {
+	if ( ! $favicon && function_exists( 'mai_get_url' ) ) {
 		return mai_get_url() . 'assets/img/icon-256x256.png';
 	}
+
 	return $favicon;
 }
 
@@ -292,6 +295,37 @@ function mai_write_to_file( $value, $filename = '__debug' ) {
 	echo "\r\n\r\n";
 	fwrite( $handle, ob_get_clean() );
 	fclose( $handle );
+}
+
+add_action( 'plugins_loaded', 'mai_load_vendor_plugins' );
+/**
+ * Load mai-engine included plugin files.
+ * This needs to run earlier than the other files.
+ *
+ * @access private
+ *
+ * @since 2.18.0
+ *
+ * @return void
+ */
+function mai_load_vendor_plugins() {
+	$files = [];
+
+	if ( ! class_exists( 'acf_pro' ) ) {
+		$files[] = '../vendor/advanced-custom-fields/advanced-custom-fields-pro/acf';
+	}
+
+	if ( ! class_exists( 'Kirki' ) ) {
+		$files[] = '../vendor/aristath/kirki/kirki';
+	}
+
+	if ( ! $files ) {
+		return;
+	}
+
+	foreach ( $files as $file ) {
+		require_once __DIR__ . "/$file.php";
+	}
 }
 
 add_action( 'after_setup_theme', 'mai_load_files', 0 );
@@ -419,14 +453,6 @@ function mai_load_files() {
 				'admin/upgrade',
 			]
 		);
-	}
-
-	if ( ! class_exists( 'acf_pro' ) ) {
-		$files[] = '../vendor/advanced-custom-fields/advanced-custom-fields-pro/acf';
-	}
-
-	if ( ! class_exists( 'Kirki' ) ) {
-		$files[] = '../vendor/aristath/kirki/kirki';
 	}
 
 	if ( class_exists( 'FacetWP' ) ) {