+++ lib/support/wpml.php
@@ -14,13 +14,13 @@ defined( 'ABSPATH' ) || die;
 
 add_filter( 'mai_template_part_objects', 'mai_wpml_template_part_objects' );
 /**
- * Gets translated template parts.
+ * Gets translated content areas.
  *
  * @uses WPML
  *
  * @since 2.16.0
  *
- * @param array $template_parts The template parts objects.
+ * @param array $template_parts The content areas objects.
  *
  * @return array
  */